+++ src/main/java/org/whiskeysierra/helsing/hadoop/functions/StandardDeviation.java
@@ -27,8 +27,8 @@ final class StandardDeviation extends SingleArgumentNumericAggregator {
     }
 
     @Override
-    public Double getResult() {
-        return Math.sqrt(Math.abs(sumOfSquares.getResult() - average.getResult() * (double) sum.getResult()) / (double) count.getResult());
+    public Long getResult() {
+        return (long) Math.sqrt(Math.abs(sumOfSquares.getResult() - average.getResult() * (double) sum.getResult()) / (double) count.getResult());
     }
 
 }
